# bundle: Preallocate buffers for file contents.