@philipaconrad 2024-06-17
@johanfylling If you think a Golang benchmark or something would be valuable to add to this PR to help the case for merging, please LMK, and I'll allocate some time for that.

Thanks again for the helpful review comments! :smile: