@srenatus 2024-06-18
The change will improve the performance when loading large (giant) bundles...I think it might be hard to do this in a benchmark. The PR's instructions outline the procedure used to verify it...that might have to be sufficient for now, I suppose.