@philipaconrad 2024-06-18
@johanfylling @srenatus Benchmarks are up! :smile:

It took me a sec to find the right way to exercise the changes-- I wound up settling on calling `Bundle.Read()`, since that invokes all of the loader machinery behind the scenes, and allows the resulting buffers to be used directly by their intended downstream consumers.

Example CLI invocations of the new benchmarks:
```shell
go test -benchmem -run=^$ -bench ^BenchmarkTarballLoader$ github.com/open-policy-agent/opa/bundle -count=30 > tar.txt
go test -benchmem -run=^$ -bench ^BenchmarkDirectoryLoader$ github.com/open-policy-agent/opa/bundle -count=30 > dir.txt
```

<details>
<summary><b>Benchstat results for <code>BenchmarkTarballLoader</code> (N=30 runs each)</b></summary>

 - [main-tar.txt](https://github.com/user-attachments/files/15893348/main-tar.txt)
 - [pr-6818-tar.txt](https://github.com/user-attachments/files/15893349/pr-6818-tar.txt)

```
$ benchstat main-tar.txt pr-6818-tar.txt
name                    old time/op    new time/op    delta
TarballLoader/1000-8      0.00ns ±59%    0.00ns ±36%     ~     (p=0.159 n=28+30)
TarballLoader/10000-8     0.02ns ±36%    0.01ns ±19%  -69.13%  (p=0.000 n=30+28)
TarballLoader/100000-8    0.09ns ±20%    0.04ns ± 4%  -54.52%  (p=0.000 n=30+27)
TarballLoader/250000-8    0.18ns ± 6%    0.10ns ± 5%  -43.45%  (p=0.000 n=30+28)

name                    old alloc/op   new alloc/op   delta
TarballLoader/1000-8       0.00B          0.00B          ~     (all equal)
TarballLoader/10000-8      0.00B          0.00B          ~     (all equal)
TarballLoader/100000-8     0.00B          0.00B          ~     (all equal)
TarballLoader/250000-8     0.00B          0.00B          ~     (all equal)

name                    old allocs/op  new allocs/op  delta
TarballLoader/1000-8        0.00           0.00          ~     (all equal)
TarballLoader/10000-8       0.00           0.00          ~     (all equal)
TarballLoader/100000-8      0.00           0.00          ~     (all equal)
TarballLoader/250000-8      0.00           0.00          ~     (all equal)
```

</details>

<details>
<summary><b>Benchstat results for <code>BenchmarkDirectoryLoader</code> (N=30 runs each)</b></summary>

 - [main-dir.txt](https://github.com/user-attachments/files/15893352/main-dir.txt)
 - [pr-6818-dir.txt](https://github.com/user-attachments/files/15893351/pr-6818-dir.txt)


```
$ benchstat main-dir.txt pr-6818-dir.txt
name                      old time/op    new time/op    delta
DirectoryLoader/10000-8     0.01ns ±18%    0.00ns ±83%   -67.90%  (p=0.000 n=27+30)
DirectoryLoader/100000-8    0.06ns ± 8%    0.02ns ±15%   -68.65%  (p=0.000 n=29+29)
DirectoryLoader/250000-8    0.11ns ±16%    0.05ns ±13%   -55.78%  (p=0.000 n=29+27)
DirectoryLoader/500000-8    0.21ns ± 6%    0.10ns ±18%   -55.17%  (p=0.000 n=26+27)

name                      old alloc/op   new alloc/op   delta
DirectoryLoader/10000-8      0.00B          0.00B           ~     (all equal)
DirectoryLoader/100000-8     0.00B          0.00B           ~     (all equal)
DirectoryLoader/250000-8     0.00B          0.00B           ~     (all equal)
DirectoryLoader/500000-8     1.00B ± 0%     0.00B       -100.00%  (p=0.000 n=30+30)

name                      old allocs/op  new allocs/op  delta
DirectoryLoader/10000-8       0.00           0.00           ~     (all equal)
DirectoryLoader/100000-8      0.00           0.00           ~     (all equal)
DirectoryLoader/250000-8      0.00           0.00           ~     (all equal)
DirectoryLoader/500000-8      0.00           0.00           ~     (all equal)
```

</details>